+++ src/commands/index.ts
@@ -8,6 +8,7 @@ import * as stats from "./stats";
 import * as top from "./top";
 import * as settings from "./settings";
 import * as insights from "./insights";
+import * as goal from "./goal";
 import * as dev from "./dev";
 
 export type BotCommand = {
@@ -28,6 +29,7 @@ export const commands: BotCommand[] = [
     top,
     settings,
     insights,
+    goal,
     dev,
 ];